+++ packages/shared/src/index.ts
@@ -69,6 +69,7 @@ export * from "./cyber.js";
 export * from "./opex.js";
 export * from "./promptTemplates.js";
 export * from "./publicApi.js";
+export * from "./talk.js";
 export * from "./industry.js";
 export * from "./healthEcosystem.js";
 export * from "./projectContinuity.js";